Create a code repository 'moveos/moveos-verifier' and use the verifier in framework building and integration testing.
