+++ crates/rooch/src/lib.rs
@@ -13,7 +13,6 @@ use commands::init::Init;
 use rooch_types::cli::{CliResult, CommandAction};
 
 pub mod commands;
-pub mod extended_checks;
 
 #[derive(clap::Parser)]
 #[clap(author, version, about, long_about = None)]
