+++ moveos/moveos-stdlib/Cargo.toml
@@ -38,6 +38,7 @@ move-package = { workspace = true }
 move-prover = { workspace = true }
 
 moveos-types = { workspace = true }
+moveos-verifier = { workspace = true }
 
 [dev-dependencies]
 #TODO remove this dependence, MoveOS should not depend on Rooch crate
