+++ moveos/moveos-verifier/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "moveos-verifier"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+anyhow = { workspace = true }
+serde = { workspace = true }
+serde_json = { workspace = true }
+codespan-reporting = { workspace = true }
+termcolor = { workspace = true }
+
+move-core-types = { workspace = true }
+move-model = { workspace = true }
+move-binary-format = { workspace = true }
+move-package = { workspace = true }
+move-compiler = { workspace = true }
+move-command-line-common = { workspace = true }
+bcs = { workspace = true }
